requirements.txt
aioboto3==15.1.0
aiobotocore==2.24.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
bitarray==3.7.1
boto3==1.39.11
botocore==1.39.11
certifi==2025.8.3
charset-normalizer==3.4.3
ckzg==2.1.3
click==8.2.1
cytoolz==1.0.1
ecdsa==0.19.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
fastapi==0.116.1
frozenlist==1.7.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jmespath==1.0.1
lighter-sdk @ git+https://github.com/elliottech/lighter-python.git@d0009799970aad54ebb940aa3dc90cbc00028c54
mangum==0.19.0
multidict==6.6.4
parsimonious==0.10.0
propcache==0.3.2
pyasn1==0.6.1
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
rlp==4.1.0
rsa==4.9.1
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
toolz==1.0.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.0.7
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
